+++ src/frontend/src/pages/provider/home.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useMemo } from "react";
 import SPHeaderNextjs from "../../components/provider/SPHeader";
 import ProviderStatsNextjs from "../../components/provider/ProviderStats";
-import BookingRequestsNextjs from "../../components/provider/BookingRequestsNextjs";
+import BookingRequestsNextjs from "../../components/provider/BookingRequests";
 import ServiceManagementNextjs from "../../components/provider/ServiceManagement";
 import CredentialsDisplayNextjs from "../../components/provider/CredentialsDisplay";
 import BottomNavigationNextjs from "../../components/provider/BottomNavigation";